+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   flutter_svg: ^2.0.7
   page_transition: ^2.0.9
   flutter_toolbar: ^0.0.2
+  image_picker_for_web: ^3.0.2
 
 dev_dependencies:
   flutter_test:
